+++ Tests/Creature/Components/EmotionsComponentTests.cs
@@ -6,7 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Tests.Components.Creature
+namespace Tests.Creature.Components
 {
     class EmotionsComponentTests : TestBase
     {
