+++ Tests/Creature/Components/PersonalityComponentTests.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Tests.Components.Creature
+namespace Tests.Creature.Components
 {
     class PersonalityComponentTests : TestBase
     {
