+++ Tests/HouseGenerationIntegrationTests.cs
@@ -9,7 +9,7 @@ using SadConsole;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Tests.Integration
+namespace Tests
 {
     class HouseGenerationIntegrationTests
     {
