+++ Tests/Items/ItemFactoryTests.cs
@@ -6,7 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Tests.Entity
+namespace Tests.Items
 {
     class ItemFactoryTests : TestBase
     {
